Update slopes and momentum definitions.

+++ src/EUTelMillepede.cc
@@ -46,7 +46,7 @@ void EUTelMillepede::computeAlignmentToMeasurementJacobian( EUTelState &state){
 	state.print();
 	float TxLocal =  state.getSlopeX();
 	float TyLocal =  state.getSlopeY();
-	const float* localpos = state.getPosition();
+	const double* localpos = state.getPosition();
 	streamlog_out( DEBUG0 ) << "This is px/pz, py/pz (local) "<< TxLocal <<","<< TyLocal << std::endl;
 	streamlog_out( DEBUG0 ) << "Local frame position "<< *localpos<<","<<*(localpos+1)<<","<<*(localpos+2) << std::endl;
 	computeAlignmentToMeasurementJacobian( *localpos,*(localpos+1), TxLocal, TyLocal);
